Renamed a lot of things, difficulty_button.py is a new class for all difficulty buttons in difficulty view. It is a working version, i will proceed tomorrow with cleaning some stuff from Start View
