+++ gui/buttons/quit_button.py
@@ -1,6 +1,7 @@
 import arcade
 import arcade.gui
 
+
 class QuitButton(arcade.gui.UIFlatButton):
     def __init__(self):
         super(QuitButton, self).__init__(text="Quit Game", width=200)
