+++ gui/buttons/start_button.py
@@ -1,8 +1,6 @@
-import arcade
 import arcade.gui
-
-from gui.screen_configuration import ScreenConfiguration
 import gui.view_manager
+from gui.screen_configuration import ScreenConfiguration
 
 
 class StartButton(arcade.gui.UIFlatButton):
@@ -13,5 +11,5 @@ class StartButton(arcade.gui.UIFlatButton):
 
     def on_click(self, event: arcade.gui.UIOnClickEvent):
         view_manager = gui.view_manager.ViewManager()
-        view_manager.show_difficulty_view()
+        view_manager.show_difficulty_view(self.config)
         self.manager.disable()
