+++ gui/text_fields/rules.py
@@ -1,0 +1,8 @@
+import arcade
+
+
+class Rules(arcade.gui.UITextArea):
+    def __int__(self):
+        super(Rules, self)
+        self.fit_content()
+
