+++ gui/views/difficulty_view.py
@@ -2,6 +2,7 @@ import arcade
 import arcade.gui
 
 from Constants import EASY, MEDIUM, HARD
+from gui.buttons.difficulty_button import DifficultyButton
 from gui.screen_configuration import ScreenConfiguration
 import gui.view_manager
 
@@ -15,7 +16,7 @@ class DifficultyView(arcade.View):
 
         self.config = screen_config
 
-        self.scaling_x = self.config.current_x / 9
+        # self.scaling_x = self.config.current_x / 9
         self.scaling_y = self.config.current_y / 15
         # --- Required for all code that uses UI element,
         # a UIManager to handle the UI.
@@ -29,22 +30,17 @@ class DifficultyView(arcade.View):
         self.v_box = arcade.gui.UIBoxLayout()
 
         # Add the difficulty buttons
-        self.easy_button = arcade.gui.UIFlatButton(text="Easy", height=self.scaling_y,
-                                                   width=self.scaling_x)
-        self.medium_button = arcade.gui.UIFlatButton(text="Medium", height=self.scaling_y,
-                                                     width=self.scaling_x)
-        self.hard_button = arcade.gui.UIFlatButton(text="Hard", height=self.scaling_y,
-                                                   width=self.scaling_x)
+        self.easy_button = DifficultyButton(button_text="Easy", manager=self.manager, mode=EASY,
+                                            screen_config=self.config)
+        self.medium_button = DifficultyButton(button_text="Medium", manager=self.manager, mode=MEDIUM,
+                                              screen_config=self.config)
+        self.hard_button = DifficultyButton(button_text="Hard", manager=self.manager, mode=HARD,
+                                            screen_config=self.config)
 
         # Add to V_Box with spacing
         self.v_box.add(self.easy_button.with_space_around(bottom=self.scaling_y / 3))
         self.v_box.add(self.medium_button.with_space_around(bottom=self.scaling_y / 3))
         self.v_box.add(self.hard_button)
-
-        # Add events
-        self.easy_button.on_click = self.easy_button_clicked
-        self.medium_button.on_click = self.medium_button_clicked
-        self.hard_button.on_click = self.hard_button_clicked
 
         # Create a widget to hold the v_box widget, that will center the buttons
         self.manager.add(
@@ -53,19 +49,6 @@ class DifficultyView(arcade.View):
                 anchor_y="center_y",
                 child=self.v_box)
         )
-
-    def easy_button_clicked(self, event):
-        self.button_clicked(EASY)
-
-    def medium_button_clicked(self, event):
-        self.button_clicked(MEDIUM)
-
-    def hard_button_clicked(self, event):
-        self.button_clicked(HARD)
-
-    def button_clicked(self, mode):
-        self.view_manager.show_game_view(mode)
-        self.manager.disable()
 
     def on_draw(self):
         self.clear()
